--- client/components/AuthForm/styles.js
@@ -1,8 +1,0 @@
-import { makeStyles } from "@material-ui/core";
-
-export default makeStyles((theme) => ({
-  content: {
-    flexGrow: 1,
-    padding: theme.spacing(8),
-  },
-}));
