+++ client/components/AuthForms/Login/Login.js
@@ -1,5 +1,8 @@
 import React, { useState } from "react";
-import useStyles from "./stylesSignup";
+import useStyles from "../Signup/stylesSignup";
+import { useDispatch, useSelector } from "react-redux";
+import { authenticate, error } from "../../../store";
+import { useHistory, Link } from "react-router-dom";
 import {
   Typography,
   Grid,
@@ -8,13 +11,11 @@ import {
   Button,
   Avatar,
 } from "@material-ui/core";
-import { authenticate, error } from "../../store";
-import { useDispatch, useSelector } from "react-redux";
-import { useHistory, Link } from "react-router-dom";
 
-const Signup = ({ formName }) => {
+const Login = ({ formName }) => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+  const classes = useStyles();
 
   const { error } = useSelector((state) => state.auth);
 
@@ -27,22 +28,27 @@ const Signup = ({ formName }) => {
     history.push("/");
   };
 
-  const classes = useStyles();
-
   return (
     <Container component="main" maxWidth="xs">
       <div className={classes.content}>
-        <Typography component="h1" variant="h5" className={classes.signuphook}>
-          Order in seconds, delivered in 30 minutes or less.
+        <Typography
+          component="h1"
+          variant="h5"
+          style={{ fontWeight: "bolder", fontSize: "2em" }}
+        >
+          Sign In
         </Typography>
-        <Typography varaint="subtitle2" className={classes.underSignUpHook}>
-          You sign up. We deliver — fast.
+        <Typography style={{ marginBottom: "-20px" }}>
+          New here?{" "}
+          <Link to="/signup" style={{ color: "deepskyblue" }}>
+            Sign Up
+          </Link>
         </Typography>
         <form className={classes.form} onSubmit={handleSubmit}>
-          <Grid container spacing={2}>
+          <Grid container spacing={0.5}>
             <Grid item xs={12}>
               <TextField
-                variant="filled"
+                variant="outlined"
                 required
                 fullWidth
                 id="email"
@@ -71,7 +77,7 @@ const Signup = ({ formName }) => {
                 fullWidth
                 variant="contained"
               >
-                Sign Up
+                Sign In
               </Button>
             ) : (
               <Button
@@ -81,7 +87,7 @@ const Signup = ({ formName }) => {
                 variant="contained"
                 disabled
               >
-                Sign Up
+                Sign In
               </Button>
             )}
           </Grid>
@@ -91,29 +97,4 @@ const Signup = ({ formName }) => {
   );
 };
 
-export default Signup;
-
-/*  if we need this later. form for first name and last name
-<Grid item xs={12} sm={6}>
-              <TextField
-                name="firstName"
-                variant="outlined"
-                required
-                fullWidth
-                id="firstName"
-                label="First Name"
-                autoFocus
-              />
-            </Grid>
-
-            <Grid item xs={12} sm={6}>
-              <TextField
-                variant="outlined"
-                required
-                fullWidth
-                id="lastName"
-                label="Last Name"
-                name="lastName"
-              />
-            </Grid>
-            */
+export default Login;
